% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis.R
\name{hypothesis.bbcor}
\alias{hypothesis.bbcor}
\title{Hypothesis method for bbcor objects}
\usage{
\method{hypothesis}{bbcor}(hypothesis, obj, cri_level = 0.9, rope = NULL)
}
\arguments{
\item{hypothesis}{A number}
}
\description{
Hypothesis method for bbcor objects
}
