% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis.R
\name{hypothesis.BGGM}
\alias{hypothesis.BGGM}
\title{Hypothesis method for BGGM objects}
\usage{
\method{hypothesis}{BGGM}(hypothesis, obj, cri_level = 0.9, rope = NULL)
}
\description{
Hypothesis method for BGGM objects
}
