% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis.R
\name{hypothesis.data.frame}
\alias{hypothesis.data.frame}
\title{#' Hypothesis method for data.frame objects}
\usage{
\method{hypothesis}{data.frame}(hypothesis, obj, cri_level = 0.9, rope = NULL)
}
\arguments{
\item{hypothesis}{A number}
}
\description{
#' Hypothesis method for data.frame objects
}
