% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis.R
\name{hypothesis}
\alias{hypothesis}
\title{Perform a linear combination of posterior samples}
\usage{
hypothesis(obj, ...)
}
\arguments{
\item{obj}{An object of class \code{BGGM}, \code{bbcor}, or \code{data.frame}}

\item{hypothesis}{A number}
}
\value{
An object of class \code{hypothesis}
}
\description{
Perform a linear combination of posterior samples
}
\examples{
add(1, 1)
}
