% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin_comb.R
\name{lin_comb.bbcor}
\alias{lin_comb.bbcor}
\title{Hypothesis method for bbcor objects}
\usage{
lin_comb.bbcor(lin_comb, obj, cri_level = 0.9, rope = NULL)
}
\arguments{
\item{hypothesis}{A number}
}
\description{
Hypothesis method for bbcor objects
}
