% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin_comb.R
\name{lin_comb.data.frame}
\alias{lin_comb.data.frame}
\title{#' Hypothesis method for data.frame objects}
\usage{
lin_comb.data.frame(lin_comb, obj, cri_level = 0.9, rope = NULL)
}
\arguments{
\item{hypothesis}{A number}
}
\description{
#' Hypothesis method for data.frame objects
}
