% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin_comb.R
\name{lin_comb}
\alias{lin_comb}
\title{Perform a linear combination of posterior samples}
\usage{
lin_comb(lin_comb, obj, cri_level = 0.9, rope = NULL)
}
\arguments{
\item{obj}{An object of class \code{BGGM}, \code{bbcor}, or \code{data.frame}}

\item{hypothesis}{A number}
}
\value{
An object of class \code{hypothesis}
}
\description{
Perform a linear combination of posterior samples
}
\examples{
add(1, 1)
}
