% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.hypothesis}
\alias{print.hypothesis}
\title{Print formatted summary of hypothesis}
\usage{
\method{print}{hypothesis}(x, ...)
}
\arguments{
\item{x}{An object of class \code{hypothesis}}

\item{...}{Other arguments to be passed to \code{print}}
}
\value{
An object of class \code{hypothesis}
}
\description{
Print formatted summary of hypothesis
}
\examples{
add(1, 1)
}
