% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin_comb.R
\name{lin_comb.data.frame}
\alias{lin_comb.data.frame}
\title{#' Hypothesis method for data.frame objects}
\usage{
lin_comb.data.frame(lin_comb, obj, cri_level = 0.9, rope = NULL)
}
\arguments{
\item{lin_comb}{A string specifying a linear combination of variables}

\item{obj}{An object of class \code{BGGM}, \code{bbcor}, or a \code{data.frame} of posterior samples}

\item{cri_level}{The level for which a credible interval should be computed}

\item{rope}{Specify a ROPE. Optional.}
}
\description{
#' Hypothesis method for data.frame objects
}
