% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.bayeslincom}
\alias{plot.bayeslincom}
\title{Perform a linear combination of posterior samples}
\usage{
\method{plot}{bayeslincom}(
  x,
  bins = 30,
  point_col = "black",
  hist_col = "black",
  hist_fill = "gray",
  bar_col = "steelblue"
)
}
\arguments{
\item{hypothesis}{A number}

\item{obj}{An object of class \code{BGGM}, \code{bbcor}, or \code{data.frame}}
}
\value{
An object of class \code{hypothesis}
}
\description{
Perform a linear combination of posterior samples
}
\examples{
add(1, 1)
}
